Legacy imports into separate package
